requirements.txt
transformers==4.46.2
python-dotenv==1.0.1
pandas==2.0.3